+++ bika/health/catalog/patient_catalog.py
@@ -38,6 +38,7 @@ _indexes_dict = {
     'getFullname': 'FieldIndex',
     # (values from PatientIdentifiers field)
     'getPatientIdentifiers': 'KeywordIndex',
+    'getClientPatientID': 'FieldIndex',
 }
 # Defining the columns for this catalog
 _columns_list = [