+++ bika/health/upgrade/v01_01_001.py
@@ -9,6 +9,7 @@ from bika.health import logger
 from bika.health.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
+from bika.health.catalog.patient_catalog import CATALOG_PATIENT_LISTING
 
 version = '1.1.1'
 profile = 'profile-{0}:default'.format(product)
@@ -27,7 +28,8 @@ def upgrade(tool):
 
     logger.info("Upgrading {0}: {1} -> {2}".format(product, ver_from, version))
 
-    # -------- ADD YOUR STUFF HERE --------
+    ut.addIndex(CATALOG_PATIENT_LISTING, 'getClientPatientID', 'FieldIndex')
+    ut.refreshCatalogs()
 
     logger.info("{0} upgraded to version {1}".format(product, version))